requirements.txt
accelerate==0.20.3
altair==5.0.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
attrs==23.1.0
av==10.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
blis==0.7.9
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
confection==0.0.4
contexttimer==0.3.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
data==0.4
decorator==5.1.1
decord==0.6.0
deepspeed==0.9.2
distlib==0.3.6
einops==0.6.1
executing==1.2.0
fairscale==0.4.4
filelock==3.12.2
fonttools==4.40.0
fsspec==2023.6.0
ftfy==6.1.1
funcsigs==1.0.2
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.31
hjson==3.1.0
huggingface-hub==0.15.1
identify==2.5.24
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipdb==0.13.13
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kaggle==1.5.13
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
ninja==1.11.1
nodeenv==1.8.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.1
pandas==2.0.2
parameterized==0.9.0
parso==0.8.3
pathy==0.10.2
peft==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.7.0
plotly==5.15.0
portalocker==2.7.0
pre-commit==3.3.3
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.31.0
rich==13.4.2
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
streamlit==1.23.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
thinc==8.1.10
tifffile==2023.4.12
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.30.0
triton==2.0.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
virtualenv==20.23.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.48
yacs==0.1.8
zipp==3.15.0
wandb